+++ layers/CMSSchema/packages/customposts-wp/src/TypeAPIs/AbstractCustomPostTypeAPI.php
@@ -308,6 +308,20 @@ abstract class AbstractCustomPostTypeAPI extends UpstreamAbstractCustomPostTypeA
             ];
             unset($query['date-to']);
         }
+        
+        // A page can have an ancestor
+        if (isset($query['parent-id'])) {
+            $query['post_parent'] = $query['parent-id'];
+            unset($query['parent-id']);
+        }
+        if (isset($query['parent-ids'])) {
+            $query['post_parent__in'] = $query['parent-ids'];
+            unset($query['parent-ids']);
+        }
+        if (isset($query['exclude-parent-ids'])) {
+            $query['post_parent__not_in'] = $query['exclude-parent-ids'];
+            unset($query['exclude-parent-ids']);
+        }
 
         $query = App::applyFilters(
             self::HOOK_QUERY,