+++ layers/CMSSchema/packages/customposts/src/FieldResolvers/ObjectType/AbstractGenericCustomPostObjectTypeFieldResolver.php
@@ -6,17 +6,21 @@ namespace PoPCMSSchema\CustomPosts\FieldResolvers\ObjectType;
 
 use PoPCMSSchema\CustomPosts\TypeAPIs\CustomPostTypeAPIInterface;
 use PoPCMSSchema\CustomPosts\TypeHelpers\CustomPostUnionTypeHelpers;
+use PoPCMSSchema\SchemaCommons\DataLoading\ReturnTypes;
+use PoPSchema\SchemaCommons\Constants\QueryOptions;
 use PoP\ComponentModel\Feedback\ObjectTypeFieldResolutionFeedbackStore;
-use PoP\ComponentModel\FieldResolvers\ObjectType\AbstractObjectTypeFieldResolver;
+use PoP\ComponentModel\FieldResolvers\ObjectType\AbstractQueryableObjectTypeFieldResolver;
 use PoP\ComponentModel\QueryResolution\FieldDataAccessorInterface;
 use PoP\ComponentModel\Schema\SchemaTypeModifiers;
 use PoP\ComponentModel\TypeResolvers\ConcreteTypeResolverInterface;
 use PoP\ComponentModel\TypeResolvers\ObjectType\ObjectTypeResolverInterface;
+use PoP\ComponentModel\TypeResolvers\ScalarType\IntScalarTypeResolver;
 use PoP\GraphQLParser\Spec\Parser\Ast\FieldInterface;
 
-abstract class AbstractGenericCustomPostObjectTypeFieldResolver extends AbstractObjectTypeFieldResolver
+abstract class AbstractGenericCustomPostObjectTypeFieldResolver extends AbstractQueryableObjectTypeFieldResolver
 {
     private ?CustomPostTypeAPIInterface $customPostTypeAPI = null;
+    private ?IntScalarTypeResolver $intScalarTypeResolver = null;
 
     final protected function getCustomPostTypeAPI(): CustomPostTypeAPIInterface
     {
@@ -27,6 +31,15 @@ abstract class AbstractGenericCustomPostObjectTypeFieldResolver extends Abstract
         }
         return $this->customPostTypeAPI;
     }
+    final protected function getIntScalarTypeResolver(): IntScalarTypeResolver
+    {
+        if ($this->intScalarTypeResolver === null) {
+            /** @var IntScalarTypeResolver */
+            $intScalarTypeResolver = $this->instanceManager->getInstance(IntScalarTypeResolver::class);
+            $this->intScalarTypeResolver = $intScalarTypeResolver;
+        }
+        return $this->intScalarTypeResolver;
+    }
 
     /**
      * @return string[]
@@ -36,6 +49,8 @@ abstract class AbstractGenericCustomPostObjectTypeFieldResolver extends Abstract
         return [
             'parent',
             'ancestors',
+            'children',
+            'childCount',
         ];
     }
 
@@ -44,6 +59,8 @@ abstract class AbstractGenericCustomPostObjectTypeFieldResolver extends Abstract
         return match ($fieldName) {
             'parent' => CustomPostUnionTypeHelpers::getCustomPostUnionOrTargetObjectTypeResolver(),
             'ancestors' => CustomPostUnionTypeHelpers::getCustomPostUnionOrTargetObjectTypeResolver(),
+            'children' => CustomPostUnionTypeHelpers::getCustomPostUnionOrTargetObjectTypeResolver(),
+            'childCount' => $this->getIntScalarTypeResolver(),
             default => parent::getFieldTypeResolver($objectTypeResolver, $fieldName),
         };
     }
@@ -52,6 +69,8 @@ abstract class AbstractGenericCustomPostObjectTypeFieldResolver extends Abstract
     {
         return match ($fieldName) {
             'ancestors' => SchemaTypeModifiers::IS_ARRAY | SchemaTypeModifiers::IS_NON_NULLABLE_ITEMS_IN_ARRAY,
+            'childCount' => SchemaTypeModifiers::NON_NULLABLE,
+            'children' => SchemaTypeModifiers::NON_NULLABLE | SchemaTypeModifiers::IS_ARRAY | SchemaTypeModifiers::IS_NON_NULLABLE_ITEMS_IN_ARRAY,
             default => parent::getFieldTypeModifiers($objectTypeResolver, $fieldName),
         };
     }
@@ -61,6 +80,8 @@ abstract class AbstractGenericCustomPostObjectTypeFieldResolver extends Abstract
         return match ($fieldName) {
             'parent' => $this->__('Parent custom post', 'customposts'),
             'ancestors' => $this->__('List of all ancestor custom posts (parent, grandparent, etc.)', 'customposts'),
+            'children' => $this->__('Child custom posts', 'customposts'),
+            'childCount' => $this->__('Number of child custom posts', 'customposts'),
             default => parent::getFieldDescription($objectTypeResolver, $fieldName),
         };
     }
@@ -79,6 +100,19 @@ abstract class AbstractGenericCustomPostObjectTypeFieldResolver extends Abstract
             case 'ancestors':
                 /** @var array<int|string> */
                 return $customPostTypeAPI->getAncestorCustomPostIDs($customPost);
+        }
+
+        $query = array_merge(
+            $this->convertFieldArgsToFilteringQueryArgs($objectTypeResolver, $fieldDataAccessor),
+            [
+                'parent-id' => $objectTypeResolver->getID($customPost),
+            ]
+        );
+        switch ($fieldDataAccessor->getFieldName()) {
+            case 'children':
+                return $customPostTypeAPI->getCustomPosts($query, [QueryOptions::RETURN_TYPE => ReturnTypes::IDS]);
+            case 'childCount':
+                return $customPostTypeAPI->getCustomPostCount($query);
         }
 
         return parent::resolveValue($objectTypeResolver, $object, $fieldDataAccessor, $objectTypeFieldResolutionFeedbackStore);