+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Pass parent on custom post mutations (#3160)
 - Added `slugPath` field to CustomPost (comprising the slug for all ancestor posts) (#3163)
 - Added fields `GenericCustomPost.ancestors` and `Page.ancestors` (#3167)
+- Added fields `children` and `childCount` to `GenericCustomPost` (#3168)
 
 ### Improvements